fix: 환경변수 관련 설정
weather api와 spotify api 관련 토큰을 환경 변수로 설정할 수 있도록 코드 변경
