--- Server/musicforecast/src/main/java/mainproject/musicforecast/domain/spotify/SpotifyToken.java
@@ -1,27 +1,0 @@
-package mainproject.musicforecast.domain.spotify;
-
-import com.wrapper.spotify.SpotifyApi;
-import com.wrapper.spotify.exceptions.SpotifyWebApiException;
-import com.wrapper.spotify.model_objects.credentials.ClientCredentials;
-import com.wrapper.spotify.requests.authorization.client_credentials.ClientCredentialsRequest;
-
-import java.io.IOException;
-
-public class SpotifyToken {
-
-    private static final String CLIENT_ID = "";
-    private static final String CLIENT_SECRET = "";
-    private static final SpotifyApi spotifyApi = new SpotifyApi.Builder().setClientId(CLIENT_ID).setClientSecret(CLIENT_SECRET).build();
-
-    public static String accessToken() {
-        ClientCredentialsRequest clientCredentialsRequest = spotifyApi.clientCredentials().build();
-        try {
-            final ClientCredentials clientCredentials = clientCredentialsRequest.execute();
-            spotifyApi.setAccessToken(clientCredentials.getAccessToken());
-            return spotifyApi.getAccessToken();
-        } catch (IOException | SpotifyWebApiException | org.apache.hc.core5.http.ParseException e) {
-            System.out.println("Error: " + e.getMessage());
-            return "error";
-        }
-    }
-}
