+++ Server/musicforecast/src/main/resources/application.yml
@@ -13,3 +13,10 @@ jwt:
   key: ${JWT_SECRET_KEY}
   access-token-expiration-minutes: 30
   refresh-token-expiration-minutes: 420
+weather:
+  api:
+    key: ${WEATHER_API_KEY}
+spotify:
+  client:
+    id: ${SPOTIFY_CLIENT_ID}
+    secret: ${SPOTIFY_CLIENT_SECRET}
